Added simple culling to determine whether or not an Entity should be drawn at render time. Useful for devices where poly count actually matters.

+++ Core/Graphics.h
@@ -49,7 +49,7 @@ namespace Monocle
 		float rotation;
 		virtual void Update()=0;
 		virtual void Render(Entity *entity)=0;
-		virtual void GetWidthHeight(int *width, int *height)=0;
+		virtual void GetWidthHeight(float *width, float *height)=0;
 	};
 
 	//! \brief Static class responsble for rendering to the screen.
